feat: renamed util to utilities (#330)
