+++ dwcj-engine/src/main/java/org/dwcj/component/button/Button.java
@@ -15,7 +15,7 @@ import org.dwcj.component.button.sink.ButtonClickEventSink;
 import org.dwcj.component.event.EventDispatcher;
 import org.dwcj.component.event.EventListener;
 import org.dwcj.component.window.AbstractWindow;
-import org.dwcj.util.BBjFunctionalityHelper;
+import org.dwcj.utilities.BBjFunctionalityHelper;
 
 
 /**
