+++ dwcj-engine/src/main/java/org/dwcj/component/slider/Slider.java
@@ -14,8 +14,7 @@ import org.dwcj.component.TabTraversable;
 import org.dwcj.component.slider.event.SliderScrollEvent;
 import org.dwcj.component.slider.sink.SliderScrollEventSink;
 import org.dwcj.component.window.AbstractWindow;
-import org.dwcj.util.BBjFunctionalityHelper;
-
+import org.dwcj.utilities.BBjFunctionalityHelper;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
