+++ dwcj-engine/src/main/java/org/dwcj/component/textarea/TextArea.java
@@ -17,8 +17,7 @@ import org.dwcj.component.TextHighlightable;
 import org.dwcj.component.textarea.event.TextAreaModifyEvent;
 import org.dwcj.component.textarea.sink.TextAreaModifyEventSink;
 import org.dwcj.component.window.AbstractWindow;
-import org.dwcj.util.BBjFunctionalityHelper;
-
+import org.dwcj.utilities.BBjFunctionalityHelper;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
