+++ dwcj-engine/src/main/java/org/dwcj/utilities/Assets.java
@@ -1,4 +1,4 @@
-package org.dwcj.util;
+package org.dwcj.utilities;
 
 import java.io.BufferedReader;
 import java.io.IOException;
